+++ src/main/java/nordmods/uselessreptile/client/gui/URDragonScreen.java
@@ -128,7 +128,7 @@ public class URDragonScreen<T extends AbstractContainerMenu> extends AbstractCon
         EntityRenderer<? super LivingEntity, ?> renderer = RenderUtil.getEntityRenderer(entity);
         LivingEntityRenderState state = (LivingEntityRenderState) renderer.createRenderState(entity, tickDelta);
         state.nameTag = null;
-        state.setStateData(URStateDataTypes.PASSENGERS_SHOULD_RENDER_TO_CLIENT, state.getStateData(URStateDataTypes.PASSENGERS_SHOULD_RENDER_TO_CLIENT).stream().map(val -> false).toList());
+        state.setStateData(URStateDataTypes.PASSENGERS_SHOULD_RENDER_TO_CLIENT, state.getStateDataOptional(URStateDataTypes.PASSENGERS_SHOULD_RENDER_TO_CLIENT).stream().map(val -> false).toList());
         state.setStateData(ClientStateDataTypes.LIGHT, LightTexture.FULL_BRIGHT);
 
         Quaternionf rot = new Quaternionf();