+++ src/main/java/nordmods/uselessreptile/datagen/data/mod/URDragonModelProvider.java
@@ -48,41 +48,41 @@ public class URDragonModelProvider implements DataProvider {
     }
 
     protected void addSpawnEntries() {
-        addWyvern("jeb_");
-        addWyvern("green");
-        addWyvern("brown");
+        addWyvern("jeb_", true);
+        addWyvern("green", false);
+        addWyvern("brown", false);
 
-        addMoleclaw("black");
-        addMoleclaw("brown");
-        addMoleclaw("grey");
-        addMoleclaw("albino");
+        addMoleclaw("black", false);
+        addMoleclaw("brown", false);
+        addMoleclaw("grey", false);
+        addMoleclaw("albino", false);
 
-        addLightningChaser("blue");
-        addLightningChaser("grey");
-        addLightningChaser("brown");
-        addLightningChaser("purple");
+        addLightningChaser("blue", false);
+        addLightningChaser("grey", false);
+        addLightningChaser("brown", false);
+        addLightningChaser("purple", false);
 
-        addRiverPikehorn("green");
-        addRiverPikehorn("dark_green");
-        addRiverPikehorn("blue");
-        addRiverPikehorn("dark_blue");
-        addRiverPikehorn("purple");
-        addRiverPikehorn("dark_purple");
-        addRiverPikehorn("teal");
-        addRiverPikehorn("dark_teal");
+        addRiverPikehorn("green", false);
+        addRiverPikehorn("dark_green", false);
+        addRiverPikehorn("blue", false);
+        addRiverPikehorn("dark_blue", false);
+        addRiverPikehorn("purple", false);
+        addRiverPikehorn("dark_purple", false);
+        addRiverPikehorn("teal", false);
+        addRiverPikehorn("dark_teal", false);
 
-        addMagmamuncher("netherrack");
-        addMagmamuncher("magma");
+        addMagmamuncher("netherrack", false);
+        addMagmamuncher("magma", false);
     }
 
-    protected ModelData getModelData(Identifier id, String variant, boolean cull) {
-        Identifier texture = Identifier.fromNamespaceAndPath(id.getNamespace(), "textures/entity/" + id.getPath() + "/" + variant + ".png");
+    protected ModelData getModelData(Identifier id, String variant, boolean cull, boolean animatedTexture) {
+        Identifier texture = Identifier.fromNamespaceAndPath(id.getNamespace(), "textures/" + (animatedTexture ? "animated_textures/" : "") + "entity/" + id.getPath() + "/" + variant + ".png");
         Identifier model = Identifier.fromNamespaceAndPath(id.getNamespace(), "biscuit_roll/models/entity/" + id.getPath() + "/" + id.getPath() + ".geo.json");
         Identifier animation = Identifier.fromNamespaceAndPath(id.getNamespace(), "biscuit_roll/animations/entity/" + id.getPath() + "/" + id.getPath() + ".animation.json");
         return new ModelData(texture, model, animation, cull, false);
     }
 
-    protected void addWyvern(String variant) {
+    protected void addWyvern(String variant, boolean animatedTexture) {
         List<DragonModelData.Sound> sounds = new ArrayList<>();
         sounds.add(new DragonModelData.Sound("step", URSoundEvent.WYVERN_STEP.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("woosh", URSoundEvent.DRAGON_WOOSH.location(), Optional.of(2f), Optional.empty(), Optional.empty()));
@@ -92,10 +92,10 @@ public class URDragonModelProvider implements DataProvider {
         sounds.add(new DragonModelData.Sound("idle", URSoundEvent.WYVERN_AMBIENT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("hurt", URSoundEvent.WYVERN_HURT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("death", URSoundEvent.WYVERN_DEATH.location(), Optional.empty(), Optional.empty(), Optional.empty()));
-        addEntry(UREntities.WYVERN_ENTITY, variant, Optional.of(sounds), true);
+        addEntry(UREntities.WYVERN_ENTITY, variant, Optional.of(sounds), true, animatedTexture);
     }
 
-    protected void addMoleclaw(String variant) {
+    protected void addMoleclaw(String variant, boolean animatedTexture) {
         List<DragonModelData.Sound> sounds = new ArrayList<>();
         sounds.add(new DragonModelData.Sound("step", URSoundEvent.DRAGON_STEP.location(), Optional.empty(), Optional.of(0.7f), Optional.empty()));
         sounds.add(new DragonModelData.Sound("attack_strong", URSoundEvent.MOLECLAW_STRONG_ATTACK.location(), Optional.empty(), Optional.empty(), Optional.empty()));
@@ -104,10 +104,10 @@ public class URDragonModelProvider implements DataProvider {
         sounds.add(new DragonModelData.Sound("hurt", URSoundEvent.MOLECLAW_HURT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("death", URSoundEvent.MOLECLAW_DEATH.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("panic", URSoundEvent.MOLECLAW_PANICKING.location(), Optional.empty(), Optional.empty(), Optional.empty()));
-        addEntry(UREntities.MOLECLAW_ENTITY, variant, Optional.of(sounds), false);
+        addEntry(UREntities.MOLECLAW_ENTITY, variant, Optional.of(sounds), false, animatedTexture);
     }
 
-    protected void addRiverPikehorn(String variant) {
+    protected void addRiverPikehorn(String variant, boolean animatedTexture) {
         List<DragonModelData.Sound> sounds = new ArrayList<>();
         sounds.add(new DragonModelData.Sound("step", SoundEvents.CHICKEN_STEP.location(), Optional.of(0.5f), Optional.of(0.8f), Optional.empty()));
         sounds.add(new DragonModelData.Sound("woosh", URSoundEvent.DRAGON_WOOSH.location(), Optional.of(0.7f), Optional.of(1.2f), Optional.empty()));
@@ -116,10 +116,10 @@ public class URDragonModelProvider implements DataProvider {
         sounds.add(new DragonModelData.Sound("idle", URSoundEvent.PIKEHORN_AMBIENT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("hurt", URSoundEvent.PIKEHORN_HURT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("death", URSoundEvent.PIKEHORN_DEATH.location(), Optional.empty(), Optional.empty(), Optional.empty()));
-        addEntry(UREntities.RIVER_PIKEHORN_ENTITY, variant, Optional.of(sounds), true);
+        addEntry(UREntities.RIVER_PIKEHORN_ENTITY, variant, Optional.of(sounds), true, animatedTexture);
     }
 
-    protected void addLightningChaser(String variant) {
+    protected void addLightningChaser(String variant, boolean animatedTexture) {
         List<DragonModelData.Sound> sounds = new ArrayList<>();
         sounds.add(new DragonModelData.Sound("step", URSoundEvent.DRAGON_STEP.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("woosh", URSoundEvent.DRAGON_WOOSH.location(), Optional.of(2f), Optional.empty(), Optional.empty()));
@@ -131,10 +131,10 @@ public class URDragonModelProvider implements DataProvider {
         sounds.add(new DragonModelData.Sound("accept_challenge", URSoundEvent.LIGHTNING_CHASER_ACCEPT_CHALLENGE.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("hurt", URSoundEvent.LIGHTNING_CHASER_HURT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("death", URSoundEvent.LIGHTNING_CHASER_DEATH.location(), Optional.empty(), Optional.empty(), Optional.empty()));
-        addEntry(UREntities.LIGHTNING_CHASER_ENTITY, variant, Optional.of(sounds), true);
+        addEntry(UREntities.LIGHTNING_CHASER_ENTITY, variant, Optional.of(sounds), true, animatedTexture);
     }
 
-    protected void addMagmamuncher(String variant) {
+    protected void addMagmamuncher(String variant, boolean animatedTexture) {
         List<DragonModelData.Sound> sounds = new ArrayList<>();
         sounds.add(new DragonModelData.Sound("step", SoundEvents.NETHERRACK_STEP.location(), Optional.of(0.25f), Optional.of(0.8f), Optional.empty()));
         sounds.add(new DragonModelData.Sound("bite", URSoundEvent.MAGMAMUNCHER_BITE.location(), Optional.empty(), Optional.empty(), Optional.empty()));
@@ -142,19 +142,19 @@ public class URDragonModelProvider implements DataProvider {
         sounds.add(new DragonModelData.Sound("hurt", URSoundEvent.MAGMAMUNCHER_HURT.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("death", URSoundEvent.MAGMAMUNCHER_DEATH.location(), Optional.empty(), Optional.empty(), Optional.empty()));
         sounds.add(new DragonModelData.Sound("apply_fire_resistance", SoundEvents.FIRECHARGE_USE.location(), Optional.empty(), Optional.empty(), Optional.empty()));
-        addEntry(UREntities.MAGMAMUNCHER_ENTITY, variant, Optional.of(sounds), true);
+        addEntry(UREntities.MAGMAMUNCHER_ENTITY, variant, Optional.of(sounds), true, animatedTexture);
     }
 
     protected void addEntry(Identifier id, DragonModelData variant) {
         holder.add(new Tuple<>(id, variant));
     }
 
-    protected void addEntry(Identifier dragonId, String variant, Optional<List<DragonModelData.Sound>> sounds, boolean cull) {
-        addEntry(id(dragonId, variant), new DragonModelData(getModelData(dragonId, variant, cull), sounds));
+    protected void addEntry(Identifier dragonId, String variant, Optional<List<DragonModelData.Sound>> sounds, boolean cull, boolean animatedTexture) {
+        addEntry(id(dragonId, variant), new DragonModelData(getModelData(dragonId, variant, cull, animatedTexture), sounds));
     }
 
-    protected void addEntry(EntityType<? extends Entity> entityType, String variant, Optional<List<DragonModelData.Sound>> sounds, boolean cull) {
-        addEntry(EntityType.getKey(entityType), variant, sounds, cull);
+    protected void addEntry(EntityType<? extends Entity> entityType, String variant, Optional<List<DragonModelData.Sound>> sounds, boolean cull, boolean animatedTexture) {
+        addEntry(EntityType.getKey(entityType), variant, sounds, cull, animatedTexture);
     }
 
     protected Identifier id(Identifier dragonId, String variant) {